+++ src/components/DevDock/DevDock.js
@@ -12,12 +12,17 @@ const devDock = React.createClass({
         return {
             subject: "",
             predicate: "",
-            object: ""
+            object: "",
+            color: "",
         }
     },
     render: function(){
         return <div id="devDock">
             <div>
+                <div>
+                    <label htmlFor="edgeColor">edgeColor:</label>
+                    <input id="edgeColor" value={this.state.color} onChange={(e) => {this.setState({color: e.target.value})}} />
+                </div>
                 <div>
                     <label htmlFor="subject">Subject:</label>
                     <input id="subject" value={this.state.subject} onChange={(e) => {this.setState({subject: e.target.value})}} />
@@ -31,7 +36,7 @@ const devDock = React.createClass({
                     <input id="object" value={this.state.object} onChange={(e) => {this.setState({object: e.target.value})}} />
                 </div>
             </div>
-            <button onClick={() => {this.props.addTriplet(this.state.subject, this.state.predicate, this.state.object)}}>Add Triplet</button>
+            <button onClick={() => {this.props.addTriplet(this.state.subject, this.state.predicate, this.state.object, this.state.color || "black")}}>Add Triplet</button>
             <button onClick={() => {this.props.removeNodeHash(this.state.subject)}}>Remove Node (from subject input)</button>
         </div>
     }
@@ -42,8 +47,8 @@ const mapStateToProps = state => ({
 });
 
 const mapDispatchToProps = dispatch => ({
-    addTriplet: (subjectHash, predicateType, objectHash) => dispatch(addTriplet(
-        {hash: subjectHash}, {type: predicateType}, {hash: objectHash})),
+    addTriplet: (subjectHash, predicateType, objectHash, color) => dispatch(addTriplet(
+        {hash: subjectHash}, {type: predicateType, color}, {hash: objectHash})),
     removeNodeHash: nodeHash => dispatch(removeNodeHash(nodeHash)),
 });