+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+ - Added predicate edge color option (with marker color change)
  - Added dynamic node size
  - Added remove node by hash
  - Stopped Graph Jumping (handleDisconnect to false)